requirements.txt
azure-functions
azure-functions-durable
azure-identity
azure-storage-blob
requests
